8,9 All Done

+++ DTO/ProductDTO.cs
@@ -4,8 +4,8 @@
     {
         public string ProductName { get; set; }
 
-        public int Price { get; set; }
+        public decimal Price { get; set; }
 
-        public int Stock { get; set; }
+        public decimal Stock { get; set; }
     }
 }
